+++ src/services/policyAssignment/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazurePolicyAssignmentInput!]!) {
+  addazurePolicyAssignment(input: $input, upsert: true) {
+    numUids
+  }
+}`;